refactor: move each individual form into a single fragment component

+++ app/components/Form/ProjectManagerForm.tsx
@@ -1,12 +1,11 @@
 import { JSONSchema7, JSONSchema7Definition } from "json-schema";
 import React, { useRef, useMemo } from "react";
 import { graphql, useFragment } from "react-relay";
-import { ProjectManagerForm_query$key } from "__generated__/ProjectManagerForm_query.graphql";
+import { ProjectManagerForm_managerFormChange$key } from "__generated__/ProjectManagerForm_managerFormChange.graphql";
 import FormBase from "./FormBase";
 import projectManagerSchema from "data/jsonSchemaForm/projectManagerSchema";
 import { ValidatingFormProps } from "./Interfaces/FormValidationTypes";
 import Grid from "@button-inc/bcgov-theme/Grid";
-import FormBorder from "lib/theme/components/FormBorder";
 import { Button } from "@button-inc/bcgov-theme";
 import validateFormWithErrors from "lib/helpers/validateFormWithErrors";
 import useAddManagerToRevisionMutation from "mutations/Manager/addManagerToRevision";
@@ -14,9 +13,14 @@ import useDeleteManagerFromRevisionMutation from "mutations/Manager/deleteManage
 import { mutation as updateFormChangeMutation } from "mutations/FormChange/updateFormChange";
 import useDebouncedMutation from "mutations/useDebouncedMutation";
 import EmptyObjectFieldTemplate from "lib/theme/EmptyObjectFieldTemplate";
+import FieldLabel from "lib/theme/widgets/FieldLabel";
 
 interface Props extends ValidatingFormProps {
-  query: ProjectManagerForm_query$key;
+  managerFormChange: ProjectManagerForm_managerFormChange$key;
+  allCifUsers: any;
+  projectId: number;
+  projectRevisionId: string;
+  projectRevisionRowId: number;
 }
 
 const uiSchema = {
@@ -33,43 +37,25 @@ const uiSchema = {
 
 const ProjectManagerForm: React.FC<Props> = (props) => {
   const formRefs = useRef({});
-  const { allCifUsers, projectRevision } = useFragment(
+  const { allCifUsers, projectId, projectRevisionId, projectRevisionRowId } =
+    props;
+
+  const change = useFragment(
     graphql`
-      fragment ProjectManagerForm_query on Query {
-        projectRevision(id: $projectRevision) {
+      fragment ProjectManagerForm_managerFormChange on ManagerFormChangesByLabelCompositeReturn {
+        projectManagerLabel {
           id
           rowId
-          managerFormChanges: projectManagerFormChangesByLabel {
-            edges {
-              node {
-                projectManagerLabel {
-                  id
-                  rowId
-                  label
-                }
-                formChange {
-                  id
-                  newFormData
-                }
-              }
-            }
-          }
-          projectFormChange {
-            formDataRecordId
-          }
+          label
         }
-        allCifUsers {
-          edges {
-            node {
-              rowId
-              firstName
-              lastName
-            }
-          }
+        formChange {
+          id
+          operation
+          newFormData
         }
       }
     `,
-    props.query
+    props.managerFormChange
   );
 
   // Dynamically build the schema from the list of cif_users
@@ -95,22 +81,28 @@ const ProjectManagerForm: React.FC<Props> = (props) => {
   const addManager = (data: any) => {
     applyAddManagerToRevision({
       variables: {
-        projectRevision: projectRevision.id,
-        projectRevisionId: projectRevision.rowId,
+        projectRevision: projectRevisionId,
+        projectRevisionId: projectRevisionRowId,
         newFormData: data,
       },
-    });
-  };
-
-  // Delete a manager from the project revision
-  const [discardFormChange] = useDeleteManagerFromRevisionMutation();
-  const deleteManager = (id: string) => {
-    discardFormChange({
-      variables: {
-        input: {
-          id: id,
+      optimisticResponse: {
+        createFormChange: {
+          query: {
+            projectRevision: {
+              managerFormChanges: {
+                edges: {
+                  change: {
+                    projectManagerLabel: {},
+                    formChange: {
+                      id: "new",
+                      newFormData: data,
+                    },
+                  },
+                },
+              },
+            },
+          },
         },
-        projectRevision: projectRevision.id,
       },
     });
   };
@@ -118,20 +110,62 @@ const ProjectManagerForm: React.FC<Props> = (props) => {
   const [applyUpdateFormChangeMutation] = useDebouncedMutation(
     updateFormChangeMutation
   );
+
+  // Delete a manager from the project revision
+  const [discardFormChange, discardInFlight] =
+    useDeleteManagerFromRevisionMutation();
+  const deleteManager = (id: string) => {
+    if (change.formChange.operation === "CREATE")
+      discardFormChange({
+        variables: {
+          input: {
+            id: id,
+          },
+          projectRevision: projectRevisionId,
+        },
+        onError: (error) => {
+          console.log(error);
+        },
+      });
+    else
+      applyUpdateFormChangeMutation({
+        variables: {
+          input: {
+            id: id,
+            formChangePatch: {
+              operation: "ARCHIVE",
+            },
+          },
+        },
+        optimisticResponse: {
+          updateFormChange: {
+            formChange: {
+              id: id,
+              newFormData: {},
+            },
+          },
+        },
+        onError: (error) => {
+          console.log(error);
+        },
+        debounceKey: id,
+      });
+  };
+
   // Update an existing project_manager form change if it exists, otherwise create one
   const createOrUpdateFormChange = (
     formChangeId: string,
     labelId: number,
-    change: any
+    formChange: any
   ) => {
     const data = {
-      ...change,
+      ...formChange,
       projectManagerLabelId: labelId,
-      projectId: projectRevision.projectFormChange.formDataRecordId,
+      projectId: projectId,
     };
 
     // If a form_change already exists, and the payload contains a cifUserId update it
-    if (formChangeId && change.cifUserId) {
+    if (formChangeId && formChange?.cifUserId) {
       applyUpdateFormChangeMutation({
         variables: {
           input: {
@@ -152,11 +186,11 @@ const ProjectManagerForm: React.FC<Props> = (props) => {
         debounceKey: formChangeId,
       });
       // If a form_change does not exist, and the payload contains a cifUserId create a form_change record
-    } else if (change.cifUserId) {
+    } else if (formChange?.cifUserId) {
       addManager(data);
     }
     // If a form_change exists, and the payload does not contain a cifUserId delete it
-    else if (formChangeId && !change.cifUserId) {
+    else if (formChangeId && !change.cifUserId && !discardInFlight) {
       deleteManager(formChangeId);
     }
   };
@@ -172,63 +206,48 @@ const ProjectManagerForm: React.FC<Props> = (props) => {
 
   return (
     <>
-      <Grid cols={10} align="center">
+      <React.Fragment key={change.projectManagerLabel.id}>
         <Grid.Row>
-          <Grid.Col span={10}>
-            <FormBorder title="Project Managers">
-              {projectRevision.managerFormChanges.edges.map(({ node }) => (
-                <React.Fragment key={node.projectManagerLabel.id}>
-                  <Grid.Row>
-                    <label>{node.projectManagerLabel.label}</label>
-                  </Grid.Row>
-                  <Grid.Row>
-                    <Grid.Col span={6}>
-                      <FormBase
-                        id={`form-manager-${node.projectManagerLabel.label}`}
-                        idPrefix={`form-${node.projectManagerLabel.id}`}
-                        ref={(el) =>
-                          (formRefs.current[node.projectManagerLabel.id] = el)
-                        }
-                        formData={node.formChange?.newFormData}
-                        onChange={(change) => {
-                          createOrUpdateFormChange(
-                            node.formChange?.id,
-                            node.projectManagerLabel.rowId,
-                            change.formData
-                          );
-                        }}
-                        schema={managerSchema}
-                        uiSchema={uiSchema}
-                        ObjectFieldTemplate={EmptyObjectFieldTemplate}
-                      />
-                    </Grid.Col>
-                    <Grid.Col span={4}>
-                      <Button
-                        variant="secondary"
-                        size="small"
-                        onClick={() => deleteManager(node.formChange?.id)}
-                      >
-                        Clear
-                      </Button>
-                    </Grid.Col>
-                  </Grid.Row>
-                </React.Fragment>
-              ))}
-            </FormBorder>
+          <FieldLabel
+            label={change.projectManagerLabel.label}
+            required={false}
+            htmlFor={change.projectManagerLabel.id}
+            uiSchema={uiSchema}
+          />
+        </Grid.Row>
+        <Grid.Row>
+          <Grid.Col span={6}>
+            <FormBase
+              id={`form-manager-${change.projectManagerLabel.label}`}
+              idPrefix={`form-${change.projectManagerLabel.id}`}
+              ref={(el) =>
+                (formRefs.current[change.projectManagerLabel.id] = el)
+              }
+              formData={change.formChange?.newFormData}
+              onChange={(data) => {
+                createOrUpdateFormChange(
+                  change.formChange?.id,
+                  change.projectManagerLabel.rowId,
+                  data.formData
+                );
+              }}
+              schema={managerSchema}
+              uiSchema={uiSchema}
+              ObjectFieldTemplate={EmptyObjectFieldTemplate}
+            />
+          </Grid.Col>
+          <Grid.Col span={4}>
+            <Button
+              disabled={discardInFlight || !change.formChange?.id}
+              variant="secondary"
+              size="small"
+              onClick={() => deleteManager(change.formChange?.id)}
+            >
+              Clear
+            </Button>
           </Grid.Col>
         </Grid.Row>
-      </Grid>
-      <style jsx>{`
-        :global(button.pg-button) {
-          margin-left: 0.4em;
-          margin-right: 0em;
-        }
-        :global(.right-aligned-column) {
-          display: flex;
-          justify-content: flex-end;
-          align-items: flex-start;
-        }
-      `}</style>
+      </React.Fragment>
     </>
   );
 };
